Fix for maps, and styles - wrong context!
