+++ app/src/main/java/pl/llp/aircasting/screens/new_session/choose_location/ChooseLocationViewMvcImpl.kt
@@ -142,7 +142,7 @@ class ChooseLocationViewMvcImpl : BaseObservableViewMvc<ChooseLocationViewMvc.Li
     }
 
     override fun onMapReady(googleMap: GoogleMap) {
-        styleGoogleMap(mMap!!, context)
+        mMapFragment?.context?.let { styleGoogleMap(mMap!!, it) }
 
         mMap = googleMap
         setZoomPreferences()
